+++ spec/active_interaction/filters/object_filter_spec.rb
@@ -132,7 +132,7 @@ describe ActiveInteraction::ObjectFilter, :filter do
       it 'raises an error' do
         expect do
           result
-        end.to raise_error ActiveInteraction::InvalidClassError
+        end.to raise_error ActiveInteraction::InvalidNameError
       end
     end